pyproject.toml
[project]
name = "phobos"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "icalendar>=7.0.1",
    "playwright>=1.58.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
